Improve icons
Encircled both icons, changed ToneButton to use a JavaScript based path animation rather than animating scale. This was done so that the ends of the wave form could be rounded, and it is a JS animation because Edge does not support CSS or SMIL animation for the path.
Add role="button" to IconButton
Remove errant console.log